requirements.txt
Django<1.12
djangorestframework<3.7
django-registration-2.2
